modify type for created_at to float64 from int64

+++ models/article_content.go
@@ -22,7 +22,7 @@ type ArticleContent struct {
 	Body string `json:"body,omitempty"`
 
 	// created at
-	CreatedAt int64 `json:"created_at,omitempty"`
+	CreatedAt float64 `json:"created_at,omitempty"`
 
 	// eye catch url
 	EyeCatchURL string `json:"eye_catch_url,omitempty"`
